start working on displaying  the files in the files area
