+++ middleware/files_mw.js
@@ -1,0 +1,74 @@
+const fs = require('fs');
+const xlsx = require('xlsx');
+const multer = require('multer');
+
+
+const storage = multer.diskStorage({
+
+    destination: function (req, file, cb) {
+        cb(null, 'uploads/');
+    },
+    filename: function (req, file, cb) {
+        cb(null, Date.now()+ '-' + file.originalname);
+    }
+});
+
+const upload = multer({ storage : storage });
+async function file_con (req,res,next){
+    if(!req.file)
+        return res.status(400).json({error: 'No File Uploaded.'})
+    const filepate = req.file.path;
+    console.log(req.file.originalname)
+    console.log(filepate)
+    if(!filepate.endsWith('.xlsx')){
+        fs.unlinkSync(filepate);
+        return res.status(400).json({error: 'Invalid File Format.'});
+    }
+
+    res.pageVal = {
+        mssg:"file uploaded successfully!",
+        isUploaded:true
+    }
+    next();
+}
+function readFile(req,res,next){
+    const workbook = xlsx.readFile("uploads/1707666602703-xlsx file.xlsx");
+    const worksheet = workbook.Sheets[workbook.SheetNames[0]];
+    let response = xlsx.utils.sheet_to_json(worksheet);
+    console.log(response)
+    res.response = response;
+    next();
+
+}
+function getFiles(req,res,next){
+    const dirPath = "./uploads";
+    let response;
+    fs.readdir(dirPath, (err, files) => {
+        if (err) {
+            console.error('Error reading directory:', err);
+            return;
+        }
+        response = files;
+        console.log(response)
+    })
+    //console.log(response);
+    const workbook = xlsx.readFile("uploads/1707666602703-xlsx file.xlsx");
+    const worksheet = workbook.Sheets[workbook.SheetNames[0]];
+    response = xlsx.utils.sheet_to_json(worksheet);
+    console.log(response)
+    res.response = response;
+    next();
+
+}
+
+
+
+module.exports = {
+    fs:fs,
+    xlsx:xlsx,
+    multer:multer,
+    upload:upload,
+    file_con:file_con,
+    readFile:readFile,
+    getFiles:getFiles
+}
